provider/kubernetes: Move translation of resources out of adaptor
